# 8293695: Implement isInfinite intrinsic for RISC-V