RISC-V 64 intrinsic for isInfinite follows the logic of x86 intrinsic (introduced by 8285868). This patch adds C2 match for IsInfinite nodes. Existing test is modified to run on RISC-V and passes on both release and fastdebug builds. Benchmark results are below:

before:
Benchmark                              Mode  Cnt   Score   Error Units
DoubleClassCheck.testIsInfiniteBranch  avgt   15  43.547 ± 6.843 ns/op
DoubleClassCheck.testIsInfiniteCMov    avgt   15  16.301 ± 1.386 ns/op
DoubleClassCheck.testIsInfiniteStore   avgt   15  16.230 ± 1.477 ns/op
FloatClassCheck.testIsInfiniteBranch   avgt   15  38.774 ± 3.572 ns/op
FloatClassCheck.testIsInfiniteCMov     avgt   15  15.064 ± 1.310 ns/op
FloatClassCheck.testIsInfiniteStore    avgt   15  14.967 ± 1.298 ns/op

after:
Benchmark                              Mode  Cnt   Score    Error Units
DoubleClassCheck.testIsInfiniteBranch  avgt   15  39.987 ±  6.179 ns/op
DoubleClassCheck.testIsInfiniteCMov    avgt   15  13.477 ±  1.159 ns/op
DoubleClassCheck.testIsInfiniteStore   avgt   15   9.607 ±  0.834 ns/op
FloatClassCheck.testIsInfiniteBranch   avgt   15  36.265 ±  3.168 ns/op
FloatClassCheck.testIsInfiniteCMov     avgt   15  13.230 ±  1.100 ns/op
FloatClassCheck.testIsInfiniteStore    avgt   15   9.492 ±  0.807 ns/op

According to 8285868 discussion, isNaN and isFinite methods intrinsification using the same approach might be not beneficial. I'm going to investigate it for RISC-V and propose methods intrinsification as part of further work in case it's profitable.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8293695](https://bugs.openjdk.org/browse/JDK-8293695): Implement isInfinite intrinsic for RISC-V


### Reviewers
 * [Yadong Wang](https://openjdk.org/census#yadongwang) (@yadongw - Author)
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**)
 * @bell-sw (no known github.com user name / role)
 * [Dmitry Samersoff](https://openjdk.org/census#dsamersoff) (@dsamersoff - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/10253/head:pull/10253` \
`$ git checkout pull/10253`

Update a local copy of the PR: \
`$ git checkout pull/10253` \
`$ git pull https://git.openjdk.org/jdk pull/10253/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 10253`

View PR using the GUI difftool: \
`$ git pr show -t 10253`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/10253.diff">https://git.openjdk.org/jdk/pull/10253.diff</a>

</details>